extended-chatbot/SAM/generatePptx/requirements.txt
python-pptx
boto3==1.28.57